+++ lib/hephaestus_oban/migrations/postgres.ex
@@ -13,7 +13,7 @@ defmodule HephaestusOban.Migrations.Postgres do
   use Ecto.Migration
 
   @initial_version 1
-  @current_version 3
+  @current_version 4
   @default_prefix "public"
 
   @valid_prefix ~r/^[a-z_][a-z0-9_]*$/