+++ priv/repo/migrations/20260413000000_update_id_types_for_business_key.exs
@@ -1,0 +1,13 @@
+defmodule HephaestusOban.TestRepo.Migrations.UpdateIdTypesForBusinessKey do
+  use Ecto.Migration
+
+  def up do
+    HephaestusEcto.Migration.up()
+    HephaestusOban.Migration.up()
+  end
+
+  def down do
+    HephaestusOban.Migration.down()
+    HephaestusEcto.Migration.down()
+  end
+end